Backend/requirements.txt
# Core FastAPI
fastapi>=0.109.0
uvicorn[standard]>=0.27.0

# Security
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6

# Database
aiosqlite>=0.19.0

# Validation
pydantic>=2.5.0
pydantic-settings>=2.1.0
email-validator>=2.1.0

# LangChain AI - using latest versions with pre-built wheels
langchain>=0.1.0
langchain-openai>=0.1.0
openai>=1.0.0

# Middleware
slowapi>=0.1.9
